♻️ refactor: 폼컴포넌트 분리, 상수 분리
