+++ app/_constants/address.ts
@@ -1,27 +1,57 @@
 export const ADDRESS = [
-  "서울시 종로구",
-  "서울시 중구",
-  "서울시 용산구",
-  "서울시 성동구",
-  "서울시 광진구",
-  "서울시 동대문구",
-  "서울시 중랑구",
-  "서울시 성북구",
+  "서울시 강남구",
+  "서울시 강동구",
   "서울시 강북구",
-  "서울시 도봉구",
-  "서울시 노원구",
-  "서울시 은평구",
-  "서울시 서대문구",
-  "서울시 마포구",
-  "서울시 양천구",
   "서울시 강서구",
+  "서울시 관악구",
+  "서울시 광진구",
   "서울시 구로구",
   "서울시 금천구",
-  "서울시 영등포구",
+  "서울시 노원구",
+  "서울시 도봉구",
+  "서울시 동대문구",
   "서울시 동작구",
-  "서울시 관악구",
+  "서울시 마포구",
+  "서울시 서대문구",
   "서울시 서초구",
-  "서울시 강남구",
+  "서울시 성동구",
+  "서울시 성북구",
   "서울시 송파구",
-  "서울시 강동구",
-].sort();
+  "서울시 양천구",
+  "서울시 영등포구",
+  "서울시 용산구",
+  "서울시 은평구",
+  "서울시 종로구",
+  "서울시 중구",
+  "서울시 중랑구",
+] as const;
+
+export type AddressType = (typeof ADDRESS)[number];
+
+export const FORMATTED_ADDRESS: { label: string; value: AddressType }[] = [
+  { label: "서울시 강남구", value: "서울시 강남구" },
+  { label: "서울시 강동구", value: "서울시 강동구" },
+  { label: "서울시 강북구", value: "서울시 강북구" },
+  { label: "서울시 강서구", value: "서울시 강서구" },
+  { label: "서울시 관악구", value: "서울시 관악구" },
+  { label: "서울시 광진구", value: "서울시 광진구" },
+  { label: "서울시 구로구", value: "서울시 구로구" },
+  { label: "서울시 금천구", value: "서울시 금천구" },
+  { label: "서울시 노원구", value: "서울시 노원구" },
+  { label: "서울시 도봉구", value: "서울시 도봉구" },
+  { label: "서울시 동대문구", value: "서울시 동대문구" },
+  { label: "서울시 동작구", value: "서울시 동작구" },
+  { label: "서울시 마포구", value: "서울시 마포구" },
+  { label: "서울시 서대문구", value: "서울시 서대문구" },
+  { label: "서울시 서초구", value: "서울시 서초구" },
+  { label: "서울시 성동구", value: "서울시 성동구" },
+  { label: "서울시 성북구", value: "서울시 성북구" },
+  { label: "서울시 송파구", value: "서울시 송파구" },
+  { label: "서울시 양천구", value: "서울시 양천구" },
+  { label: "서울시 영등포구", value: "서울시 영등포구" },
+  { label: "서울시 용산구", value: "서울시 용산구" },
+  { label: "서울시 은평구", value: "서울시 은평구" },
+  { label: "서울시 종로구", value: "서울시 종로구" },
+  { label: "서울시 중구", value: "서울시 중구" },
+  { label: "서울시 중랑구", value: "서울시 중랑구" },
+];
